Remove keepalived
The DRBD/keepalived based HA subsystem was deprecated as of Chef Server
12.9, and officially reached end of life on 2019-03-31. It has been
disabled in Chef Server 13.

This blocks off usage of the ha commands in chef-server-ctl, and
prevents reconfigure from supporting the ha topology.

There's a lot more work to do to complete cleanup, but this removes
key functionality in preparation for further work.

Signed-off-by: Mark Anderson <mark@chef.io>